requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
cachetools==5.3.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cryptography==40.0.2
cycler==0.11.0
dataclasses-json==0.5.7
DateTime==5.1
decorator==5.1.1
entrypoints==0.4
extra-streamlit-components==0.1.56
fake2db==0.5.4
Faker==18.11.1
favicon==0.7.0
fonttools==4.39.4
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.31
greenlet==2.0.2
htbuilder==0.6.1
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
install==1.3.5
Jinja2==3.1.2
jsonschema==4.17.3
kiwisolver==1.4.4
langchain==0.0.166
lxml==4.9.2
Markdown==3.4.3
markdown-it-py==2.2.0
markdownlit==0.0.7
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.1
mdurl==0.1.2
more-itertools==9.1.0
multidict==6.0.4
Nuitka==1.6.5
numpy==1.24.3
openapi-schema-pydantic==1.2.4
ordered-set==4.1.0
packaging==23.1
pandas==2.0.1
pandoc==2.3
pdfkit==1.0.0
pdfminer==20191125
pdfminer.six==20221105
Pillow==9.5.0
plotly==5.14.1
plumbum==1.8.1
protobuf==3.20.3
pyarrow==12.0.0
pycparser==2.21
pycryptodome==3.17
pydantic==1.10.7
pydeck==0.8.1b0
Pygments==2.15.1
pymdown-extensions==9.11
Pympler==1.0.1
pypandoc==1.11
pyparsing==3.0.9
pypng==0.20220715.0
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.4.2
requests==2.30.0
rich==13.3.5
six==1.16.0
smmap==5.0.0
soupsieve==2.4.1
SQLAlchemy==1.4.0
st-annotated-text==4.0.0
st-pages==0.4.1
streamlit==1.23.1
streamlit-antd-components==0.1.6
streamlit-authenticator==0.1.5
streamlit-camera-input-live==0.2.0
streamlit-card==0.0.4
streamlit-cookies-manager==0.2.0
streamlit-embedcode==0.1.2
streamlit-extras==0.2.7
streamlit-faker==0.0.2
streamlit-image-coordinates==0.1.3
streamlit-keyup==0.2.0
streamlit-login-auth-ui==0.2.0
streamlit-lottie==0.0.3
streamlit-option-menu==0.3.2
streamlit-pydantic==0.6.0
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==1.0.2
tenacity==8.2.2
toml==0.10.2
toolz==0.12.0
tornado==6.3.1
trycourier==4.4.0
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
urllib3==2.0.2
validators==0.20.0
watchdog==3.0.0
xlrd==2.0.1
yarl==1.9.2
zipp==3.15.0
zstandard==0.21.0
